Initial commit: Introduce HD44780-compatible display library for AVR MCUs
Introducing a HD44780 compatible library for AVR MCUs.
This is based on [1], with further optimizations and additional support
for OLED-0010 controllers for OLED displays.

The only "caveat" is that in order to save space we're using bitfields,
so pin grouping is sort of "hardcoded", refer to LCD_LIB.h for more information.

[1] https://www.youtube.com/watch?v=SMVevawvbqE&list=PLmX2_OhrubDDvx8ZnrQOitygvdq6sIwj0

Signed-off-by: Daniel Vásquez <danielgusvt@gmail.com>